requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
weaviate-client==3.15.6
openai==0.27.8
python-magic==0.4.27
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
transformers==4.30.2
torch==2.0.1
soundfile==0.12.1
librosa==0.10.0
numpy==1.24.3
pydub==0.25.1
SpeechRecognition==3.10.0
